refactor: extract scoped OIDC claim wiring into PhysicalTenantOidcProviders
The two build helpers that sourced per-registration claim converters and
identity claims from the physical-tenant configuration lived in the OIDC
@Configuration class and duplicated the same forAllPhysicalTenants walk.
Move that domain logic into a static factory in the pt package, next to
PhysicalTenantAuthConfigurations where the physical-tenant provider
knowledge already lives, and share a single providersByRegistrationId
helper between both derivations. The configuration class now only wires
the beans and stays thin.
